# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice System API
  description: |
    REST API for the invoice reconciliation platform (CSV/XLSX accounting
    imports, reconciliation runs, notification emails, audit trail, users,
    company and application settings).

    ## Versioning

    * **Stable, documented namespace:** `/api/v1` (this document).
    * **Compatibility namespace:** `/api` resolves to the *same* handlers
      and remains supported so existing clients (including the bundled SPA
      frontend) keep working unchanged. No business logic is duplicated;
      both prefixes are registered against identical view functions.

    ## Response contract

    Every endpoint returns JSON. Success payloads use:

    ```json
    { "success": true, "message": "...", "data": { } }
    ```

    Error payloads always use the stable error envelope:

    ```json
    {
      "success": false,
      "message": "Human readable reason",
      "status": 400,
      "code": "BAD_REQUEST"
    }
    ```

    See the `Error` schema and the `codes` values in the components section.

    ## Correlation

    Every response carries an `X-Request-Id` header. A client-supplied value
    (alphanumeric/dot/dash, up to 64 chars) is echoed back; otherwise a
    generated UUID is returned. Use it to correlate client requests with
    server logs.

    ## Deprecation

    Superseded endpoints (marked `deprecated: true`) keep working unchanged,
...
    major version.

    ## Money and dates

...
  version: "1.0.0"
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /settings/application:
    get:
      tags: [Settings]
      operationId: getApplicationSettings
      summary: Return the safe public subset of application settings.
      responses:
        "200":
          description: Settings payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [Settings]
      operationId: updateApplicationSettings
      summary: Update application settings (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated settings payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /settings/application/all:
    get:
      tags: [Settings]
      operationId: getAllApplicationSettings
      summary: Return every setting including internal ones (admin only).
      responses:
        "200":
          description: Full settings payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /settings/user:
    get:
      tags: [Settings]
      operationId: getUserSettings
      summary: Return the authenticated user's own preferences.
      responses:
        "200":
          description: User preferences.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [Settings]
      operationId: updateUserSettings
      summary: Update the authenticated user's own preferences.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated preferences.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessBody" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------------------------------
  # Imports
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token issued by POST /auth/login. Include as
        `Authorization: Bearer <token>`.
...
  schemas:
    SuccessBody:
      type: object
      required: [success]
      properties:
        success: { type: boolean, const: true }
        message:
          type: string
          description: Optional human-readable text (not required for the contract).
        data:
          description: >-
            Contract payload field. Present on endpoints that return data;
            omitted only for signal-only actions (logout, delete, password
            change). When present it is always an object or array, never a raw
            scalar.
          nullable: true
    Error:
      type: object
      required: [success, message, status, code]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        status: { type: integer }
        code:
          type: string
          description: >-
            Stable machine-readable error code. `APP_ERROR` is a reserved
            internal fallback and must never be returned for the documented
            statuses below.
          enum:
            - BAD_REQUEST
            - INVALID_CREDENTIALS
            - UNAUTHORIZED
            - TOKEN_REQUIRED
            - TOKEN_EXPIRED
            - INVALID_TOKEN
            - TOKEN_REVOKED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VALIDATION_ERROR
            - RATE_LIMITED
            - DATABASE_ERROR
            - INTERNAL_ERROR
            - BAD_GATEWAY
            - SERVICE_UNAVAILABLE
            - MISSING_FIELD
...
  responses:
...
    BadRequest:
      description: Invalid request.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Missing, expired, revoked or invalid token.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but role insufficient.
      headers: *reqid_headers
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...